chore: add jest tests
