+++ __mocks__/mcpServerIndex.js
@@ -1,0 +1,5 @@
+class Server {
+  setRequestHandler() {}
+  connect() {}
+}
+module.exports = { Server };
