+++ __mocks__/stdio.js
@@ -1,0 +1,2 @@
+class StdioServerTransport {}
+module.exports = { StdioServerTransport };
