+++ __mocks__/types.js
@@ -1,0 +1,7 @@
+module.exports = {
+  CallToolRequestSchema: {},
+  ListToolsRequestSchema: {},
+  ListPromptsRequestSchema: {},
+  GetPromptRequestSchema: {},
+  GetPromptResultSchema: {}
+};
