+++ tsconfig.json
@@ -6,6 +6,7 @@
       "outDir": "./build",
       "rootDir": "./src",
       "strict": true,
+      "isolatedModules": true,
       "esModuleInterop": true,
       "skipLibCheck": true,
       "forceConsistentCasingInFileNames": true
